Fix cyclic dependency of middleware import
