+++ agent-manager-service/clients/openchoreosvc/client/client.go
@@ -28,8 +28,8 @@ import (
 
 	"github.com/wso2/agent-manager/agent-manager-service/clients/openchoreosvc/gen"
 	"github.com/wso2/agent-manager/agent-manager-service/clients/requests"
-	"github.com/wso2/agent-manager/agent-manager-service/middleware"
 	"github.com/wso2/agent-manager/agent-manager-service/models"
+	"github.com/wso2/agent-manager/agent-manager-service/orgctx"
 )
 
 // HeaderImpersonateOrg carries the org UUID the call is performed on behalf
@@ -214,7 +214,7 @@ func NewOpenChoreoClient(cfg *Config) (OpenChoreoClient, error) {
 	// organization being operated on. Absent on contexts that don't originate
 	// from an authenticated API request (e.g. background reconcilers).
 	orgEditor := func(ctx context.Context, req *http.Request) error {
-		if org, ok := middleware.GetResolvedOrg(ctx); ok && org.OUID != "" {
+		if org, ok := orgctx.GetResolvedOrg(ctx); ok && org.OUID != "" {
 			req.Header.Set(HeaderImpersonateOrg, org.OUID)
 		}
 		return nil
